requirements.txt
absl-py==0.11.0
appdirs==1.4.3
asgiref==3.3.1
astunparse==1.6.3
attrs==20.3.0
biwrap==0.1.6
black==19.10b0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.6
easydict==1.9
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
image==1.5.33
imageio==2.9.0
importlib-metadata==3.4.0
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.1
msgpack==0.6.2
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.5
pathspec==0.8.1
pep517==0.8.2
Pillow==6.1.0
progress==1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.1.2
redis==3.4.1
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.2.2
rsa==4.7
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==0.12.1
tensorflow-estimator==2.4.0
tensorflow-gpu==0.12.1
tensorflow-plot==0.3.2
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
validators==0.18.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
